Hello World! with this commit trunk will finally be making use of the gorgeous stuff sirius_black last implemented. we want to make sure everything related to this feature is working as intended so for now its being released under a optional define, after the features stability is confirmed the define will be removed along with the previous processings of the feature. By enabling BETA_THREAD_TEST in /src/config/core.h all your mysql logs, query_sql and query_logsql script functions will be handled by a different thread, therefore any slow queries won't have any effect (e.g. slow down) on the game server. Everyone is welcome on helping us test and debug the features, be aware however I DO NOT RECOMMEND YOU TO ENABLE THIS ON PRODUCTION SERVERS AS OF NOW (which is why while we're all testing it still is a optional feature).
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@16588 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/common/thread.c
@@ -271,7 +271,7 @@ int rathread_get_tid(){
 	#ifdef WIN32
 		return (int)GetCurrentThreadId();
 	#else
-		return (int)pthread_self();
+		return (intptr_t)pthread_self();
 	#endif
 	
 #endif
